DRY up createVideoOrAudioEl
