+++ src/utils/media-utils.js
@@ -480,6 +480,29 @@ export function createBasisTexture(url) {
   });
 }
 
+const isIOS = AFRAME.utils.device.isIOS();
+
+/**
+ * Create video element to be used as a texture.
+ *
+ * @param {string} src - Url to a video file.
+ * @returns {Element} Video element.
+ */
+export function createVideoOrAudioEl(type) {
+  const el = document.createElement(type);
+  el.setAttribute("playsinline", "");
+  el.setAttribute("webkit-playsinline", "");
+  // iOS Safari requires the autoplay attribute, or it won't play the video at all.
+  el.autoplay = true;
+  // iOS Safari will not play videos without user interaction. We mute the video so that it can autoplay and then
+  // allow the user to unmute it with an interaction in the unmute-video-button component.
+  el.muted = isIOS;
+  el.preload = "auto";
+  el.crossOrigin = "anonymous";
+
+  return el;
+}
+
 export function addMeshScaleAnimation(mesh, initialScale, onComplete) {
   const step = (function() {
     const lastValue = {};
